Build device class dynamically
