+++ config/config.json
@@ -3,10 +3,13 @@
     "#device_server_name": "PogoTestDevice/test", 
     "#device_name": "test/test_device/0",
     "#device_class_path": "tango_classes/PogoTestDevice.py",
+    "#device_class_type": "PogoClass",
     "£device_server_name": "PytangoTestDevice/test", 
     "£device_name": "test/test_device/0",
     "£device_class_path": "tango_classes/PytangoTestDevice.py",
+    "£device_class_type": "PyTangoClass",
     "device_server_name": "BuildTestDevice/test", 
     "device_name": "test/test_device/0",
-    "device_class_path": "tango_classes/BuildTestDevice.obj"
+    "device_class_path": "tango_classes/BuildTestDevice.py",
+    "device_class_type": "BuildClass"
 }
