--- src/util.py
@@ -1,9 +1,0 @@
-import time
-
-
-def spin():
-    time.sleep(3)
-
-
-def dummy_measure_func(self):
-    return 1.23
